// academy-rnm/src/core/enums/images-rutes.ts
export enum ImagesRutes {
  settingsIcon = 'assets/images/home-page/settingsIcon.svg',
  imageRickAndMorty = 'assets/images/home-page/imgRnM.svg',
  lenIcon = 'assets/images/home-page/len.svg',

  statusAlive = 'assets/images/home-page/status_alive.svg',
  statusDead = 'assets/images/home-page/status_dead.svg',
  statusUnknown = 'assets/images/home-page/status_unknown.svg',

  iconStatusDead = 'assets/images/modal-character-component/status-dead.svg',
  iconStatusAlive = 'assets/images/modal-character-component/status-alive.svg',
  iconStatusUnknown = 'assets/images/modal-character-component/status-unknown.svg',

  logoLogin = 'assets/images/login-page/rick_n_morty_logo.png',
  userImg = 'assets/images/user-profile/user_image.svg'
}